requirements.txt
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
certifi==2025.4.26
cffi==1.17.1
greenlet==3.2.2
minio==7.2.15
numpy==2.2.6
pandas==2.2.3
pycparser==2.22
pycryptodome==3.23.0
python-dateutil==2.9.0.post0
pytz==2025.2
six==1.17.0
SQLAlchemy==2.0.41
typing_extensions==4.13.2
tzdata==2025.2
urllib3==2.4.0
python-dotenv